UI Normalized

+++ src/main/java/Templates/Options.java
@@ -4,12 +4,16 @@
  */
 package Templates;
 
+import Classes.Tools;
+import Classes.Window;
 import com.mongodb.MongoClient;
 import com.mongodb.MongoClientURI;
 import com.mongodb.MongoException;
 import com.mongodb.client.MongoCursor;
 import com.mycompany.cuevadeana.Main;
 import com.mycompany.cuevadeana.Mongo;
+import java.util.ArrayList;
+import java.util.List;
 import javax.swing.JOptionPane;
 
 /**
@@ -30,6 +34,8 @@ public class Options extends javax.swing.JFrame {
     public Options(Main parentFrame) {
         this.parentFrame = parentFrame;
         initComponents();
+        URI.setText(parentFrame.getURIMONGO());
+        DBNAME.setText(parentFrame.getDBNAME());
         this.setResizable(false);
     }
 
@@ -68,16 +74,16 @@ public class Options extends javax.swing.JFrame {
         jPanel2Layout.setHorizontalGroup(
             jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel2Layout.createSequentialGroup()
-                .addContainerGap()
+                .addGap(17, 17, 17)
                 .addComponent(jLabel5, javax.swing.GroupLayout.DEFAULT_SIZE, 319, Short.MAX_VALUE)
-                .addGap(97, 97, 97))
+                .addGap(86, 86, 86))
         );
         jPanel2Layout.setVerticalGroup(
             jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addGroup(jPanel2Layout.createSequentialGroup()
+                .addContainerGap()
                 .addComponent(jLabel5)
-                .addContainerGap())
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
 
         javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
@@ -99,11 +105,7 @@ public class Options extends javax.swing.JFrame {
 
         jLabel2.setText("Ingrese los datos de la base de datos");
 
-        URI.setText("mongodb://localhost:27017");
-
         jLabel3.setText("Nombre de la base de datos");
-
-        DBNAME.setText("cine");
 
         SaveButton.setBackground(new java.awt.Color(255, 51, 51));
         SaveButton.setFont(new java.awt.Font("Segoe UI", 1, 18)); // NOI18N
@@ -187,13 +189,29 @@ public class Options extends javax.swing.JFrame {
 
         pack();
     }// </editor-fold>//GEN-END:initComponents
-    private void getValues() {
-        Uri = URI.getText().trim();
-        DBName = DBNAME.getText().trim();
+    private boolean getValues() {
+        List<String> errorList = new ArrayList<>();
+        String UriString = URI.getText().trim();
+        String NameString = DBNAME.getText().trim();
+        if (!Tools.validate("dbname", NameString)) {
+            errorList.add("\n" + NameString);
+        }
+        if (!Tools.validate("uri", UriString)) {
+            errorList.add("\n" + UriString);
+        }
+        if (errorList.isEmpty()) {
+            Uri = UriString;
+            DBName = NameString;
+            return true;
+        }
+        System.out.println(Tools.validate("dbname", NameString));
+        System.out.println(Tools.validate("uri", UriString));
+        Window.Message("warning", "Los datos ingresados son invalidos: " + errorList.toString().substring(1, errorList.toString().length()), "Datos invalidos");
+        return false;
     }
     private void SaveButtonMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_SaveButtonMouseClicked
         // TODO add your handling code here:
-
+        parentFrame.changeDB_Options(Uri, DBName);
     }//GEN-LAST:event_SaveButtonMouseClicked
 
     private boolean testConnection() {
@@ -209,58 +227,61 @@ public class Options extends javax.swing.JFrame {
             }
             mongoClient.close();
         } catch (MongoException e) {
-            DebugWindow.Message("danger", "\n" + e.toString(), "Error en la prueba de conexión");
+            Window.Message("danger", "\n" + e.toString(), "Error en la prueba de conexión");
         }
         return false;
     }
     private void TestConnectionButtonMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_TestConnectionButtonMouseClicked
         // TODO add your handling code here:
-        getValues();
-        boolean status = testConnection(); // Resultado de la conexion
-        DebugWindow.Message(status ? "info" : "danger", status ? "Conexión valida" : "Conexión invalida", "Prueba de conexión");
+        if (getValues()) {
+            boolean status = testConnection(); // Resultado de la conexion
+            Window.Message(status ? "info" : "danger", status ? "Conexión valida" : "Conexión invalida", "Prueba de conexión");
+        }
     }//GEN-LAST:event_TestConnectionButtonMouseClicked
 
     private void DataDefaultMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_DataDefaultMouseClicked
         // TODO add your handling code here:
-        //Bloquear boton
-        DataDefault.setEnabled(false);
-        getValues();
-        int op = JOptionPane.showConfirmDialog(parentFrame, "¿Desea cargarlos valores iniciales del programa?\nEsto eliminara los valores existentes almacenados (incluyendo la base de datos: " + DBName + ")", "Cargar valores por defecto", JOptionPane.OK_CANCEL_OPTION);
-        if (op == 0) {
-            int op1 = JOptionPane.showConfirmDialog(parentFrame,
-                    "ATENCIÓN\nSe realizaran las siguientes operaciones:\n"
-                    + "\t1. Eliminación de la base de datos: " + DBName + "\n"
-                    + "\t2. Creación de las colecciones Bill, Showtimes, Movies y Users\n"
-                    + "\t3. Creación automatica de dos usuarios de prueba de diferente cargo (Administrador y cajero) con contraseñas: 123\n"
-                    + "=====================================================================================\n"
-                    + "Todos los datos almacenados y la base de datos " + DBName + " serán elimados, al continuar con esta\n"
-                    + "operación debera reingresar las peliculas y tiempos de funciones de cada una.",
-                    "Confirmacion de carga de datos por defecto", JOptionPane.OK_CANCEL_OPTION);
-            if (op1 == 0) {
-                String textOriginal, text;
-                textOriginal = DBName + ((int) (Math.random() * 99) + 10) + "";
-                text = JOptionPane.showInputDialog(parentFrame, "Ingrese el siguiente texto: " + textOriginal, "Ingrese el texto de seguridad", JOptionPane.WARNING_MESSAGE);
-                if (textOriginal.equals(text)) {
-                    //Cargar valores por defecto
-                    Mongo mongo = new Mongo(Uri, DBName);
-                    //Eliminamos y creamos los datos
-                    mongo.dropDatabase();
-                    if (mongo.createDefaultCollections()) {
-                        this.dispose();
-                        DebugWindow.Message("info", "Datos creados", "Exitos en la creación de los datos");
+        if (getValues()) {
+            //Bloquear boton
+            DataDefault.setEnabled(false);
+            getValues();
+            int op = JOptionPane.showConfirmDialog(parentFrame, "¿Desea cargarlos valores iniciales del programa?\nEsto eliminara los valores existentes almacenados (incluyendo la base de datos: " + DBName + ")", "Cargar valores por defecto", JOptionPane.OK_CANCEL_OPTION);
+            if (op == 0) {
+                int op1 = JOptionPane.showConfirmDialog(parentFrame,
+                        "ATENCIÓN\nSe realizaran las siguientes operaciones:\n"
+                        + "\t1. Eliminación de la base de datos: " + DBName + "\n"
+                        + "\t2. Creación de las colecciones Bill, Showtimes, Movies y Users\n"
+                        + "\t3. Creación automatica de dos usuarios de prueba de diferente cargo (Administrador y cajero) con contraseñas: 123\n"
+                        + "=====================================================================================\n"
+                        + "Todos los datos almacenados y la base de datos " + DBName + " serán elimados, al continuar con esta\n"
+                        + "operación debera reingresar las peliculas y tiempos de funciones de cada una.",
+                        "Confirmacion de carga de datos por defecto", JOptionPane.OK_CANCEL_OPTION);
+                if (op1 == 0) {
+                    String textOriginal, text;
+                    textOriginal = DBName + ((int) (Math.random() * 99) + 10) + "";
+                    text = JOptionPane.showInputDialog(parentFrame, "Ingrese el siguiente texto: " + textOriginal, "Ingrese el texto de seguridad", JOptionPane.WARNING_MESSAGE);
+                    if (textOriginal.equals(text)) {
+                        //Cargar valores por defecto
+                        Mongo mongo = new Mongo(Uri, DBName);
+                        //Eliminamos y creamos los datos
+                        mongo.dropDatabase();
+                        if (mongo.createDefaultCollections()) {
+                            this.dispose();
+                            Window.Message("info", "Datos creados", "Exitos en la creación de los datos");
+                        } else {
+                            Window.Message("danger", "Error en la creacion de los datos, se recomienda  validar los datos y reintentar", "ERROR EN LA CARGA DE DATOS POR DEFECTO");
+                            DataDefault.setEnabled(true);
+                        }
                     } else {
-                        DebugWindow.Message("danger", "Error en la creacion de los datos, se recomienda  validar los datos y reintentar", "ERROR EN LA CARGA DE DATOS POR DEFECTO");
+                        Window.Message("warning", "El texto de confimación es incorrecto", "Texto incorrecto");
                         DataDefault.setEnabled(true);
                     }
                 } else {
-                    DebugWindow.Message("warning", "El texto de confimación es incorrecto", "Texto incorrecto");
                     DataDefault.setEnabled(true);
                 }
             } else {
                 DataDefault.setEnabled(true);
             }
-        } else {
-            DataDefault.setEnabled(true);
         }
     }//GEN-LAST:event_DataDefaultMouseClicked
